Changed object type to match function signature

+++ runtime/compiler/z/codegen/J9TreeEvaluator.cpp
@@ -736,7 +736,7 @@ TR::Register * caseConversionHelper(TR::Node* node, TR::CodeGenerator* cg, bool 
    TR::Instruction* cursor;
 
    const int elementSizeMask = (isCompressedString) ? 0x0 : 0x1;    // byte or halfword mask
-   const int8_t sizeOfVector = cg->machine()->getVRFSize();
+   const int32_t sizeOfVector = cg->machine()->getVRFSize();
    const bool is64 = TR::Compiler->target.is64Bit();
    uintptrj_t headerSize = TR::Compiler->om.contiguousArrayHeaderSizeInBytes();
 
@@ -932,7 +932,7 @@ inlineIntrinsicIndexOf(TR::Node * node, TR::CodeGenerator * cg, bool isLatin1)
    TR::Register* length = cg->gprClobberEvaluate(node->getChild(4));
 
 
-   const int8_t sizeOfVector = cg->machine()->getVRFSize();
+   const int32_t sizeOfVector = cg->machine()->getVRFSize();
 
    // load length isn't used after loop, size must is adjusted to become bytes left
    TR::Register* loopCounter = length;
